requirements.txt
appdirs==1.4.4
audioread==3.0.0
Authlib==1.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
dominate==2.7.0
email-validator==2.0.0.post2
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.38.0
gcloud==0.18.3
google==3.0.0
google-api-core==2.11.0
google-auth==2.18.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-datastore==2.15.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
googledrivedownloader==0.4
grpcio==1.55.0
grpcio-status==1.55.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.6.3
miditoolkit==0.1.16
mido==1.2.10
more-itertools==9.0.0
msgpack==1.0.5
music21==8.1.0
muspy==0.5.0
numba==0.56.4
numpy==1.23.5
oauth2client==4.1.3
packaging==23.0
Pillow==9.4.0
pooch==1.6.0
postgres==4.0
pretty-midi==0.2.9
proto-plus==1.22.2
protobuf==4.23.1
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
pyFluidSynth==1.3.2
pygame==2.1.2
PyJWT==2.7.0
pyparsing==3.0.9
pypianoroll==1.0.4
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.0
six==1.16.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
SQLAlchemy==2.0.15
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
visitor==0.1.3
webcolors==1.12
Werkzeug==2.2.3
WTForms==3.0.1
